requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
boto3==1.16.57
botocore==1.19.57
dj-database-url==0.5.0
Django==3.1.5
django-background-tasks==1.2.5
django-cleanup==5.1.0
django-compat==1.0.15
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-storages==1.11.1
feedparser==6.0.2
gunicorn==20.0.4
jmespath==0.10.0
lxml==4.6.2
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.5
s3transfer==0.3.4
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
